Reorganized request tests. Fixed double callback bug.
